Update special chars on mapper

+++ server/wfnews-persistence/src/main/java/ca/bc/gov/nrs/wfnews/persistence/v1/dao/mybatis/PublishedIncidentDaoImpl.java
@@ -198,10 +198,15 @@ public class PublishedIncidentDaoImpl extends BaseDao implements
 		try {
 
 			Map<String, Object> parameters = new HashMap<String, Object>();
+			
 			Integer offset = null;
+			
 			int totalRowCount = this.publishedIncidentMapper.selectCount(parameters);
-			pageNumber = pageNumber==null?Integer.valueOf(0):pageNumber;
-			if(pageRowCount != null) { offset = Integer.valueOf((pageNumber.intValue()-1)*pageRowCount.intValue()); }
+			pageNumber = pageNumber == null ? 0 : pageNumber;
+
+			if (pageRowCount != null) { 
+				offset = Integer.valueOf((pageNumber.intValue() - 1) * pageRowCount.intValue());
+			}
 			//avoid jdbc exception for offset when pageNumber is 0
 			if (offset != null && offset < 0) offset = 0;
